Bug fix: Boiler unable to update fuel
In some cases, such as in certain modpacks which modifies fuel performance, the fuelTime variable could be above 0 but below the required amountPerTick, and in that case the boiler thinks it doesn't need another fuel piece, which would cause it to cool itself down, unable to pull another fuel item. This tiny fix would be a workaround for those scenarios.